# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "token": {
            "maxLength": 128,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "AcceptInviteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MfaCodeRequest": {
        "properties": {
          "code": {
            "pattern": "^\\s*\\d{3}\\s?\\d{3}\\s*$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "MfaCodeRequest",
        "type": "object"
      },
...
      "SessionState": {
        "description": "Where a signed-in session stands; the web app routes on ``stage``.",
        "examples": [
          {
            "active_tenant": {
              "name": "Acme Industries",
              "slug": "acme",
              "tenant_id": "01928f6e-\u2026"
            },
            "permissions": {
              "platform.users.read": [
                "tenant"
              ]
            },
            "stage": "ready",
            "tenants": [
              {
                "name": "Acme Industries",
                "slug": "acme",
                "tenant_id": "01928f6e-\u2026"
              }
            ],
            "user": {
              "display_name": "Acme Admin",
              "email": "admin@acme.test",
              "id": "01928f6e-\u2026",
              "mfa_enabled": true
            }
          }
        ],
        "properties": {
          "active_tenant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "permissions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Effective permissions and the scope types they apply in",
            "title": "Permissions",
            "type": "object"
          },
          "stage": {
            "$ref": "#/components/schemas/AuthStage"
          },
          "tenants": {
            "items": {
              "$ref": "#/components/schemas/TenantSummary"
            },
            "title": "Tenants",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "stage",
          "user",
          "active_tenant",
          "tenants"
        ],
        "title": "SessionState",
        "type": "object"
      },
...
      "TokenRequest": {
        "properties": {
          "token": {
            "maxLength": 128,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TokenRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pickwise API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/email/verify": {
      "post": {
        "operationId": "verify_email_v1_auth_email_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/invites/accept": {
      "post": {
        "operationId": "accept_invite_v1_auth_invites_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invite",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/invites/{token}": {
      "get": {
        "operationId": "inspect_invite_v1_auth_invites__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Invite",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/mfa/confirm": {
      "post": {
        "operationId": "confirm_mfa_setup_v1_auth_mfa_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Mfa Setup",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/disable": {
      "post": {
        "operationId": "disable_mfa_v1_auth_mfa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Mfa",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/recovery-codes": {
      "post": {
        "operationId": "regenerate_recovery_codes_v1_auth_mfa_recovery_codes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Recovery Codes",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/mfa/verify": {
      "post": {
        "operationId": "verify_mfa_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Mfa",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/change": {
      "post": {
        "operationId": "change_password_v1_auth_password_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/forgot": {
      "post": {
        "description": "Always 202, whether or not the email has an account.",
        "operationId": "forgot_password_v1_auth_password_forgot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Forgot Password V1 Auth Password Forgot Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "operationId": "reset_password_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sso/callback": {
      "get": {
        "operationId": "sso_callback_v1_auth_sso_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "maxLength": 256,
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sso Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sso/discover": {
      "get": {
        "description": "Which single sign-on options cover this email's domain (ids only).",
        "operationId": "sso_discover_v1_auth_sso_discover_get",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "maxLength": 254,
              "title": "Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SsoDiscovery"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sso Discover",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sso/start": {
      "get": {
        "operationId": "sso_start_v1_auth_sso_start_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sso_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sso Config Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "login_hint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Login Hint"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sso Start",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/session/tenant": {
      "post": {
        "operationId": "switch_tenant_v1_session_tenant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Tenant",
        "tags": [
          "auth",
          "me"
        ]
      }
    },
...
  }
}